+++ src/Elchristo/Calendar/Model/Event/AbstractCalendarEvent.php
@@ -154,7 +154,7 @@ abstract class AbstractCalendarEvent implements CalendarEventInterface
     /**
      * @return array
      */
-    public function getOptions()
+    public function getOptions(): array
     {
         return $this->options;
     }