+++ src/Elchristo/Calendar/Model/Event/DefaultCalendarEvent.php
@@ -5,6 +5,7 @@ namespace Elchristo\Calendar\Model\Event;
 use Elchristo\Calendar\Model\Event\AbstractCalendarEvent;
 use Elchristo\Calendar\Model\Event\ColoredEventInterface;
 use Elchristo\Calendar\Model\Event\ColoredEventTrait;
+use Elchristo\Calendar\Service\Color\ColorStrategyAwareTrait;
 
 /**
  * Default calendar event
@@ -12,4 +13,5 @@ use Elchristo\Calendar\Model\Event\ColoredEventTrait;
 class DefaultCalendarEvent extends AbstractCalendarEvent implements ColoredEventInterface
 {
     use ColoredEventTrait;
+    use ColorStrategyAwareTrait;
 }